+++ javascript/package.json
@@ -90,7 +90,7 @@
     "prettier": "3.1.0",
     "puppeteer": "^22.11.1",
     "serve-handler": "^6.1.5",
-    "ts-mocha": "^10.0.0",
+    "ts-mocha": "^11.1.0",
     "ts-node": "^10.9.1",
     "typedoc": "^0.23.22",
     "typescript": "^4.9.4"
